#58 프로필 선택 bottom sheet 유저 정보 API 적용
